% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISgrad.R
\name{CE.mixed.grad}
\alias{CE.mixed.grad}
\title{Calculate p-value of HC by cross-entropy based importance sampling method with mix Gussian gradient proposal family}
\usage{
CE.mixed.grad(
  q,
  K,
  ro = 0.01,
  N0 = 10^4,
  N1 = 10^4,
  B = 1,
  k0 = 1,
  k1 = NA,
  thre = FALSE,
  idx = 1,
  theta = 1
)
}
\arguments{
\item{q}{quantile or observed value of HC statistic, should be a scalar}

\item{K}{dimension of HC, i.e. the total number of studies}

\item{ro}{related to the heavy tailed proposal density, 0.01 by default}

\item{N0}{sampling size for automatically selecting the proposal density}

\item{N1}{sampling size of each epoch for estimating p-value, and the total size is N1*B}

\item{B}{number of epoches}

\item{k0}{search range starts from the k0th smallest p-value, default value is 1.}

\item{k1}{search range ends at the k1th smallest p-value, default value is K.}

\item{thre}{whether using thresholding supremum domain, Boolean variable equals False by default.}

\item{idx}{hyper parameter of gradient degree to specify the Gaussian mixture family, 1 by default.}

\item{theta}{initial parameter value of the proposal density family, 1 by default.}
}
\value{
The right-tail probability of the null distribution of HC statistic at the given quantile, i.e. p-value.
}
\description{
Calculate p-value of HC by cross-entropy based importance sampling method with mix Gussian gradient proposal family
}
\examples{
pset=runif(10)
hcstat <-HCstat(pset,k0=1)
CE.mixed.grad(q=hcstat, K=10)
}
