% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISprop.R
\name{CE.mixed.prop}
\alias{CE.mixed.prop}
\title{Calculate p-value of HC by cross-entropy based importance sampling method with mix Gussian proportion proposal family}
\usage{
CE.mixed.prop(
  q,
  K,
  ro = 0.01,
  N0 = 10^4,
  N1 = 10^4,
  B = 1,
  k0 = 1,
  k1 = NA,
  thre = T,
  prop = 0.2,
  theta = 1
)
}
\arguments{
\item{q}{\itemize{
\item quantile or observed value of HC statistic, should be a scalar
}}

\item{K}{\itemize{
\item dimension of HC, i.e. the total number of studies
}}

\item{ro}{related to the heavy tailed proposal density, 0.01 by default}

\item{N0}{sampling size for automatically selecting the proposal density}

\item{N1}{sampling size of each epoch for estimating p-value, and the total size is N1*B}

\item{B}{number of epoches}

\item{k0}{search range starts from the k0th smallest p-value, default value is 1.}

\item{k1}{search range ends at the k1th smallest p-value, default value is K.}

\item{thre}{whether using thresholding supremum domain, Boolean variable equals True by default.}

\item{prop}{hyper parameter of proportion to specify the Gaussian mixture family, 0.2 by default.}

\item{theta}{initial parameter value of the proposal density family, 1 by default.}
}
\value{
The right-tail probability of the null distribution of HC statistic at the given quantile, i.e. p-value.
}
\description{
Calculate p-value of HC by cross-entropy based importance sampling method with mix Gussian proportion proposal family
}
\examples{
pset=runif(10)
hcstat <-HCstat(pset,k0=1,thre=TRUE)
CE.mixed.prop(q=hcstat, K=10)
}
