% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCflibs.R
\docType{data}
\name{HC_flibs}
\alias{HC_flibs}
\title{Pre-constructed spline functions of FHC.}
\format{
Spline functions for FHC of K ranging from 2 to 20000
}
\usage{
HC_flibs
}
\description{
Include function sets for FHC to predict the p-value given observed statistic.
}
\examples{
 data(HC_flibs)
}
\keyword{datasets}
