% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCflibs_q.R
\docType{data}
\name{HC_flibs_q}
\alias{HC_flibs_q}
\title{Pre-constructed spline functions of FHC.}
\format{
Spline functions for FHC of K ranging from 2 to 20000
}
\usage{
HC_flibs_q
}
\description{
Include function sets for FHC to predict the threshold given significance level.
}
\examples{
 data(HC_flibs_q)
}
\keyword{datasets}
