% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCstat.R
\name{HCstat}
\alias{HCstat}
\title{Calculate HC statitics given a set of individual p-values}
\usage{
HCstat(pset, k0 = 1, k1 = NA, thre = F)
}
\arguments{
\item{pset}{a vector of p-values from individual studies}

\item{k0}{search range starts from the k0th smallest p-value, default value is 1.}

\item{k1}{search range ends at the k1th smallest p-value, default value is K.}

\item{thre}{whether using thresholding supremum domain, Boolean variable equals False by default.}
}
\value{
HC statitic
}
\description{
Calculate HC statitics given a set of individual p-values
}
\examples{
pset=runif(10)
hcstat <-HCstat(pset,k0=1)

}
