% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{hybrid}
\alias{hybrid}
\title{Hybrid method of computing p-value for HC of arbitrary variation}
\usage{
hybrid(q, K, k0 = 1, k1 = NA, thre = FALSE, N = 10^6)
}
\arguments{
\item{q}{quantile or observed value of HC statistic, should be a scalar}

\item{K}{dimension of HC, i.e. the total number of studies}

\item{k0}{search range starts from the k0th smallest p-value, default value is 1.}

\item{k1}{search range ends at the k1th smallest p-value, default value is K.}

\item{thre}{whether using thresholding supremum domain, Boolean variable equals False by default.}

\item{N}{Samping size when sampling method is recommended, default value is 10^6.}
}
\value{
The right-tail probability of the null distribution of HC statistic at the given quantile, i.e. p-value.
}
\description{
The function is recommended for HC of arbitrary variation except for FHC, THC, MHC, TMHC
}
\examples{
pset=runif(200)
hcstat <-HCstat(pset,k0=1,k1=80, thre=FALSE)
hybrid(q=hcstat, K=200, k0=1,k1=80, thre=FALSE,N=10^6)
}
