% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybridSpec.R
\name{hybridSpec}
\alias{hybridSpec}
\title{Hybrid method of computing p-value for HC of four popular variations}
\usage{
hybridSpec(q, K, flibs = HC_flibs, N = 10^6)
}
\arguments{
\item{q}{quantile or observed value of HC statistic, should be a scalar}

\item{K}{dimension of HC, i.e. the total number of studies}

\item{flibs}{choose the spline function data set of one specific variation of HC form HC_flibs, THC_flibs, MHC_flibs or TMHC_flibs, and the default is HC_flibs.}

\item{N}{Samping size when sampling method is recommended, default value is 10^6.}
}
\value{
The right-tail probability of the null distribution of HC statistic at the given quantile, i.e. p-value.
}
\description{
This method is recommended for four popular HC variations of arbitrary dimension k: HC of full supremum domain (FHC), HC of half truncated supremum domain (THC), HC of thresholding supremum domain (MHC) or HC of thresholding and half truncated supremum domain (TMHC)
}
\examples{
pset=runif(2001)
hcstat <-HCstat(pset,k0=1,k1=NA, thre=FALSE)
hybridSpec(q=hcstat, K=2001, flibs=HC_flibs,N=10^6)

}
