% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LS.R
\name{LiAppro_HC}
\alias{LiAppro_HC}
\title{Calculate p-value of HC by Li-Siegmund approximation}
\usage{
LiAppro_HC(q, K, k0 = 1, k1 = NA, thre = TRUE)
}
\arguments{
\item{q}{quantile or observed value of HC statistic, should be a scalar}

\item{K}{dimension of HC, i.e. the total number of studies}

\item{k0}{search range starts from the k0th smallest p-value, default value is 1.}

\item{k1}{search range ends at the k1th smallest p-value, default value is K.}

\item{thre}{whether using thresholding supremum domain, Boolean variable equals TRUE by default.}
}
\value{
The right-tail probability of the null distribution of HC statistic at the given quantile, i.e. p-value.
}
\description{
Calculate p-value of HC by Li-Siegmund approximation
}
\examples{
pset=rbeta(100,1,2)
hcstat <-HCstat(pset,k0=1,thre=TRUE)
LiAppro_HC(q=hcstat, K=length(pset))
}
