% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHCflibs.R
\docType{data}
\name{MHC_flibs}
\alias{MHC_flibs}
\title{Pre-constructed spline functions of MHC.}
\format{
Spline functions for MHC of K ranging from 20 to 20000
}
\usage{
MHC_flibs
}
\description{
Include function sets for MHC to predict the p-value given observed statistic.
}
\examples{
 data(MHC_flibs)
}
\keyword{datasets}
