% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHCflibs_q.R
\docType{data}
\name{MHC_flibs_q}
\alias{MHC_flibs_q}
\title{Pre-constructed spline functions of MHC.}
\format{
Spline functions for MHC of K ranging from 20 to 20000
}
\usage{
MHC_flibs_q
}
\description{
Include function sets for MHC to predict the threshold given significance level.
}
\examples{
 data(MHC_flibs_q)
}
\keyword{datasets}
