% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{mst}
\alias{mst}
\title{Calculate the p-value of Higher Criticism statitic given an observed value}
\usage{
mst(q, K, k0 = 1, k1 = NA, thre = FALSE)
}
\arguments{
\item{q}{quantile or observed value of HC statistic, should be a scalar}

\item{K}{dimension of HC, i.e. the total number of studies}

\item{k0}{search range starts from the k0th smallest p-value, default value is 1.}

\item{k1}{search range ends at the k1th smallest p-value, default value is K.}

\item{thre}{whether using thresholding supremum domain, Boolean variable equals False by default.}
}
\value{
The right-tail probability of the null distribution of HC statistic at the given quantile, i.e. p-value.
}
\description{
Calculate the p-value of Higher Criticism statitic given an observed value
}
\examples{
pset=runif(50)
hcstat <-HCstat(pset,k0=1)
mst(q=hcstat, K=length(pset), k0=1)
}
