% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THCflibs.R
\docType{data}
\name{THC_flibs}
\alias{THC_flibs}
\title{Pre-constructed spline functions of THC.}
\format{
Spline functions for THC of K ranging from 2 to 20000
}
\usage{
THC_flibs
}
\description{
Include function sets for THC to predict the p-value given observed statistic.
}
\examples{
 data(THC_flibs)
}
\keyword{datasets}
