% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/THCflibs_q.R
\docType{data}
\name{THC_flibs_q}
\alias{THC_flibs_q}
\title{Pre-constructed spline functions of THC.}
\format{
Spline functions for THC of K ranging from 2 to 20000
}
\usage{
THC_flibs_q
}
\description{
Include function sets for THC to predict the threshold given significance level.
}
\examples{
 data(THC_flibs_q)
}
\keyword{datasets}
