% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMHCflibs.R
\docType{data}
\name{TMHC_flibs}
\alias{TMHC_flibs}
\title{Pre-constructed spline functions of TMHC.}
\format{
Spline functions for TMHC of K ranging from 30 to 20000
}
\usage{
TMHC_flibs
}
\description{
Include function sets for TMHC to predict the p-value given observed statistic.
}
\examples{
data(TMHC_flibs)
}
\keyword{datasets}
