% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMHCflibs_q.R
\docType{data}
\name{TMHC_flibs_q}
\alias{TMHC_flibs_q}
\title{Pre-constructed spline functions of TMHC.}
\format{
Spline functions for TMHC of K ranging from 30 to 20000
}
\usage{
TMHC_flibs_q
}
\description{
Include function sets for TMHC to predict the threshold given significance level.
}
\examples{
 data(TMHC_flibs_q)
}
\keyword{datasets}
