% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFIp.R
\name{ufi.p}
\alias{ufi.p}
\title{Calculate p-value of HC by ultra-fast interpolation method}
\usage{
ufi.p(flibs = HC_flibs, K, q)
}
\arguments{
\item{flibs}{choose the spline function data set of one specific variation of HC form HC_flibs, THC_flibs, MHC_flibs or TMHC_flibs, and the default is HC_flibs.}

\item{K}{a vector of HC dimensions i.e. the total number of studies}

\item{q}{a vector of quantiles or observed values of HC statistic}
}
\value{
a vector of p-values corresponding to the input HC statistics
}
\description{
Calculate p-value of HC by ultra-fast interpolation method
}
\examples{
q=sapply(10:15,function(i) {
pset=runif(i)
return(HCstat(pset,k0=1,thre=FALSE))
})
ufi.p(flibs=HC_flibs,K=10:15,q=q)

}
