% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UFIq.R
\name{ufi.q}
\alias{ufi.q}
\title{Calculate threshold of HC by ultra-fast interpolation (UFI) method}
\usage{
ufi.q(flibs_q = HC_flibs_q, K, p)
}
\arguments{
\item{flibs_q}{choose the spline function data set of one specific variation of HC form HC_flibs_q, THC_flibs_q, MHC_flibs_q or TMHC_flibs_q, and the default is HC_flibs_q.}

\item{K}{a vector of HC dimensions i.e. the total number of studies}

\item{p}{a vector of signifiance level for each of the test}
}
\value{
a vector of thresholds of HC tests corresponding to the input significance levels
}
\description{
This method is recommended when K is no larger than 2000, for HC of full supremum domain (FHC), HC of half truncated supremum domain (THC), HC of thresholding supremum domain (MHC) or HC of thresholding and half truncated supremum domain (TMHC)
}
\examples{
ufi.q(flibs_q=HC_flibs_q,K=10:15,p=10^(seq(-3,-8)))

}
