% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BL.R
\name{XH}
\alias{XH}
\title{Barnett-Lin method of computing p-value for FHC}
\usage{
XH(q, K)
}
\arguments{
\item{q}{quantile or observed value of HC statistic, should be a scalar}

\item{K}{dimension of HC, i.e. the total number of studies}
}
\value{
The right-tail probability of the null distribution of HC statistic at the given quantile, i.e. p-value.
}
\description{
Barnett-Lin method of computing p-value for FHC
}
\examples{
pset=runif(10)
hcstat <-HCstat(pset,k0=1)
XH(q=hcstat, K=10)
}
